Add files via upload
This notebook is created to perform the data science project for my thesis data science in action of the Data Science and Society Masters program at Tilburg University. This project focussed on a multi-label classification task of different plastic litter categories found on beaches and river shores in Japan. This to try out innovative approaches to improve the field of environmental monitoring.